+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## 1.3.1 (UNRELEASED)
-- TBC
+
+- fix: ensure `rack.` internal headers are stripped in responses
+- chore: remove ancient dead Rails 2-era adapter code
+- fix: ensure ErrorApp does not mutate shared headers constant
 
 ## 1.3.0